+++ src/view/fragments/ViewRelationships.tsx
@@ -1,18 +1,20 @@
 import React, { FunctionComponent } from 'react';
 import ViewTypeFragment from './ViewTypeFragment';
 import ViewRelationshipFragment from './ViewRelationshipFragment';
-import { DataModal } from '../../models/DataModal';
+import BoxContainer from '../../components/BoxContainer';
 
-const ViewRelationships: FunctionComponent<{
-  object: DataModal;
-  objectKey: string;
-}> = ({ object, objectKey }) => {
+const ViewRelationships: FunctionComponent<any> = ({ object, template }) => {
+  console.log(template);
   return (
     <>
-      {/* how it relates */}
-      <ViewTypeFragment value={objectKey} />
-
-      <ViewRelationshipFragment relationships={object.data[objectKey]} />
+      {template.relationships.map((key: any, index: number) => (
+        <div key={index}>
+          <BoxContainer>
+            <ViewTypeFragment value={key} />
+            <ViewRelationshipFragment relationships={object.relationships} />
+          </BoxContainer>
+        </div>
+      ))}
     </>
   );
 };